move close icon to the common file
